Safari-609.1.11.1/LayoutTests/js/promise-resolve-in-non-dom-expected.txt

Test Promise.resolve in non dom directory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The promise is not fulfilled until after this function call executes.
PASS value.then is called.
PASS thisValue is value
PASS fulfilled
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
